requirements.txt
langchain
langchain_community
langchain_huggingface
flask
faiss-cpu
langchain_ollama
pypdf
rank_bm25
nltk
numpy
neo4j
requests
instructor
python-dotenv
langchain_groq
langchain_experimental
pymongo
bson
jwt